third_party/WebKit/LayoutTests/accessibility/scroll-to-make-visible-nested-2-expected.txt

Tests that scrolling to make an element visible works when the inner scroll container doesn't need to be scrolled, but the other one does.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

5000-pixel box
Target Target Target
5000-pixel box
PASS window.pageYOffset is 0
PASS container.scrollTop is 0
PASS target1.getBoundingClientRect().top is >= 5000
PASS window.innerHeight is >= target2.getBoundingClientRect().bottom
PASS target2top is target2.getBoundingClientRect().top
PASS target2top is target2.getBoundingClientRect().top
PASS successfullyParsed is true

TEST COMPLETE
